Tai28/03

+++ src/java/controller/DeleteCategoriesController.java
@@ -2,7 +2,6 @@
  * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
  * Click nbfs://nbhost/SystemFileSystem/Templates/JSP_Servlet/Servlet.java to edit this template
  */
-
 package controller;
 
 import java.io.IOException;
@@ -13,33 +12,42 @@ import jakarta.servlet.http.HttpServlet;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import model.DAOCategories;
+import model.DAOProducts;
 
 /**
  *
  * @author Admin
  */
-@WebServlet(name="DeleteCategoriesController", urlPatterns={"/DeleteCategories"})
+@WebServlet(name = "DeleteCategoriesController", urlPatterns = {"/DeleteCategories"})
 public class DeleteCategoriesController extends HttpServlet {
-   
-    /** 
-     * Processes requests for both HTTP <code>GET</code> and <code>POST</code> methods.
+
+    /**
+     * Processes requests for both HTTP <code>GET</code> and <code>POST</code>
+     * methods.
+     *
      * @param request servlet request
      * @param response servlet response
      * @throws ServletException if a servlet-specific error occurs
      * @throws IOException if an I/O error occurs
      */
     protected void processRequest(HttpServletRequest request, HttpServletResponse response)
-    throws ServletException, IOException {
+            throws ServletException, IOException {
         response.setContentType("text/html;charset=UTF-8");
-        String cid=request.getParameter("cid");
-        DAOCategories daocategories=new DAOCategories();
+        String cid = request.getParameter("cid");
+        DAOCategories daocategories = new DAOCategories();
+        DAOProducts daoproduct = new DAOProducts();
+        if (daoproduct.isCategoryExistProduct(cid)) {
+            response.sendRedirect("AdminCategories?message=Category+has+product+so+can+not+be+delete.");
+            return;
+        }
         daocategories.deleteCategories1(cid);
-        response.sendRedirect("AdminCategories");
-    } 
+        response.sendRedirect("AdminCategories?message=Delete+successfully.");
+    }
 
     // <editor-fold defaultstate="collapsed" desc="HttpServlet methods. Click on the + sign on the left to edit the code.">
-    /** 
+    /**
      * Handles the HTTP <code>GET</code> method.
+     *
      * @param request servlet request
      * @param response servlet response
      * @throws ServletException if a servlet-specific error occurs
@@ -47,12 +55,13 @@ public class DeleteCategoriesController extends HttpServlet {
      */
     @Override
     protected void doGet(HttpServletRequest request, HttpServletResponse response)
-    throws ServletException, IOException {
+            throws ServletException, IOException {
         processRequest(request, response);
-    } 
+    }
 
-    /** 
+    /**
      * Handles the HTTP <code>POST</code> method.
+     *
      * @param request servlet request
      * @param response servlet response
      * @throws ServletException if a servlet-specific error occurs
@@ -60,12 +69,13 @@ public class DeleteCategoriesController extends HttpServlet {
      */
     @Override
     protected void doPost(HttpServletRequest request, HttpServletResponse response)
-    throws ServletException, IOException {
+            throws ServletException, IOException {
         processRequest(request, response);
     }
 
-    /** 
+    /**
      * Returns a short description of the servlet.
+     *
      * @return a String containing servlet description
      */
     @Override
